New files moved in the correct folder and namespace
